*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@59 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/cspfj/problem/Variable.java
@@ -480,7 +480,7 @@ public final class Variable implements Comparable<Variable> {
     private int bestIndex;
 
     public int bestImprovment(final TabuManager tabuManager,
-            final int aspiration) {
+            final float aspiration) {
         final int bestIndex = this.bestIndex;
         if (bestIndex >= 0 && !tabuManager.isTabu(this, bestIndex)) {
             return bestIndex;
